[fish] fix make_completion for "\"
So the problem is "\" in the input will be parsed as plain "\",
instead of an escape of the following character. e.g. a directory
named "a b" will be "a\ b" on first tab, and then "a\\ b" on second
tab.

There still is a problem that if " is used to avoid using "\", the quote
itself will be escaped, returning funny result. Anyway, it's not
conflicting with my habbits, and I don't have a solution now after
spending some time, so I'll just leave it now.